+++ R/manhattan_data_preprocess.R
@@ -4,7 +4,7 @@
 #' before making a manhattan plot.
 #' It accepts a \code{data.frame}, which at bare minimum should
 #' contain a chromosome, position, and p-value.
-#' Additional options, such as chromosome color, label colum names,
+#' Additional options, such as chromosome color, label column names,
 #' and colors for specific variants, are provided here.
 #'
 #' @param x a data frame or any other extension of data frame (e.g. a tibble).
@@ -32,8 +32,8 @@
 #' with different graphic options, you have the choice to preprocess separately and
 #' then generate plots.
 #'
-#' @return a MPdata object. This object contains all the necessary info for constructing
-#'   a manhattan plot.
+#' @return a MPdata object. This object contains all the necessary components
+#' for constructing a manhattan plot.
 #'
 #' @examples
 #'
@@ -63,7 +63,7 @@ manhattan_data_preprocess.default <- function(x, ...) stop("Provide a valid data
 #' @param signif a numeric vector. Significant p-value thresholds to be drawn for
 #'   manhattan plot. At least one value should be provided. Default value is c(5e-08, 1e-5)
 #' @param pval.colname a character. Column name of \code{x} containing p.value.
-#' @param chr.colname a character. Column name of \code{x} containing chromosome number.
+#' @param chr.colname a character. Column name of \code{x} containing chromosome.
 #' @param pos.colname a character. Column name of \code{x} containing position.
 #' @param chr.order a character vector. Order of chromosomes presented in manhattan plot.
 #' @param signif.col a character vector of equal length as \code{signif}. It contains
@@ -237,7 +237,7 @@ setMethod(
       grdat, chromosome = chromosome, signif = signif, pval.colname = pval.colname,
       chr.colname = chr.colname, pos.colname = pos.colname, highlight.colname = highlight.colname, chr.order = chr.order,
       signif.col = signif.col, chr.col = chr.col, highlight.col = highlight.col, preserve.position = preserve.position, thin = thin,
-      thin.n = thin.n, chr.gap.scaling = 1, ...
+      pval.log.transform = pval.log.transform, thin.n = thin.n, chr.gap.scaling = 1, ...
     )
   }
 )